pyproject.toml
[project]
name = "flinch"
version = "0.5.0"
description = "AI content restriction consistency research tool"
readme = "README.md"
license = "MIT"
authors = [{ name = "Brad Leclerc", email = "brad@beargleindustries.com" }]
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.40.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "pyyaml>=6.0",
    "pydantic>=2.0",
    "rich>=13.0",
    "httpx>=0.27.0",
    "python-multipart>=0.0.9",
    "aiosqlite>=0.20.0",
]

[project.optional-dependencies]
multi-model = ["openai", "google-genai"]
experiment = [
    "textstat>=0.7.3",
    "scipy>=1.11.0",
    "numpy>=1.24.0",
    "matplotlib>=3.7.0",
    "krippendorff>=0.7.0",
]
dev = [
    "pytest>=8.0",
    "ruff>=0.8.0",
]

[project.scripts]
flinch = "flinch.app:main"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
